requirements.txt
fastapi
uvicorn
python-multipart
pydantic
passlib
bcrypt==4.3.0
python-jose[cryptography]
python-dotenv
pyjwt
redis
sqlalchemy
pytest
httpx
ruff
black
